moved several utils and utils related tests to core
