+++ song-core/src/test/java/org/icgc/dcc/song/core/errors/DebugTest.java
@@ -1,4 +1,4 @@
-package org.icgc.dcc.song.server.service;
+package org.icgc.dcc.song.core.errors;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -8,7 +8,7 @@ import static java.lang.Thread.currentThread;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.icgc.dcc.common.core.util.stream.Collectors.toImmutableList;
 import static org.icgc.dcc.common.core.util.stream.Streams.stream;
-import static org.icgc.dcc.song.server.utils.Debug.getCallingStackTrace;
+import static org.icgc.dcc.song.core.utils.Debug.getCallingStackTrace;
 
 @Slf4j
 public class DebugTest {
