--- song-server/src/main/java/org/icgc/dcc/song/server/exceptions/ServerError.java
@@ -1,11 +1,0 @@
-package org.icgc.dcc.song.server.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public interface ServerError {
-
-  String getErrorId();
-
-  HttpStatus getHttpStatus();
-
-}
