--- song-server/src/main/java/org/icgc/dcc/song/server/exceptions/Services.java
@@ -1,14 +1,0 @@
-package org.icgc.dcc.song.server.exceptions;
-
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor public enum Services {
-
-  ID_SERVICE("Id"),
-  UPLOAD_SERVICE("Upload");
-  @NonNull private final String name;
-
-}
