--- song-server/src/main/java/org/icgc/dcc/song/server/utils/JsonDocUtils.java
@@ -1,63 +1,0 @@
-/*
- * Copyright (c) 2017 The Ontario Institute for Cancer Research. All rights reserved.
- *
- * This program and the accompanying materials are made available under the terms of the GNU Public License v3.0.
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-package org.icgc.dcc.song.server.utils;
-
-import java.io.InputStream;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
-import com.google.common.base.Strings;
-
-import lombok.SneakyThrows;
-import lombok.val;
-
-public class JsonDocUtils {
-
-  @SneakyThrows
-  public static JsonNode getJsonNodeFromClasspath(String fileName) {
-    ObjectMapper mapper = new ObjectMapper()
-        .registerModule(new ParameterNamesModule())
-        .registerModule(new Jdk8Module())
-        .registerModule(new JavaTimeModule());
-    InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream(fileName);
-    JsonNode actualObj = mapper.readTree(is);
-    return actualObj;
-  }
-
-  @SneakyThrows
-  public static JsonNode getNode(String jsonString) {
-    ObjectMapper mapper = new ObjectMapper()
-        .registerModule(new ParameterNamesModule())
-        .registerModule(new Jdk8Module())
-        .registerModule(new JavaTimeModule());
-    return mapper.readTree(jsonString);
-  }
-
-  public static String getValue(JsonNode node, String key) {
-    val result = node.get(key).textValue();
-    if (Strings.isNullOrEmpty(result)) {
-      throw new IllegalArgumentException("No value found for " + key);
-    } else {
-      return result;
-    }
-  }
-}
