+++ song-server/src/test/java/org/icgc/dcc/song/server/service/FileServiceTest.java
@@ -21,8 +21,8 @@ package org.icgc.dcc.song.server.service;
 import lombok.val;
 import org.flywaydb.test.annotation.FlywayTest;
 import org.flywaydb.test.junit.FlywayTestExecutionListener;
+import org.icgc.dcc.song.core.utils.JsonUtils;
 import org.icgc.dcc.song.server.model.entity.File;
-import org.icgc.dcc.song.server.utils.JsonUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
